add user to balance database on creation

+++ db/migrations/20220414130433_usersbalance.js
@@ -6,7 +6,7 @@ exports.up = function (knex) {
   return knex.schema.createTable('usersbalance', (t) => {
     t.increments('id').primary().notNullable();
     t.integer('userid').notNullable().unique().references('id').inTable('users');
-    t.integer('amount').defaultTo(0);
+    t.decimal('amount').defaultTo(0);
   });
 };
